requirements.txt
asgiref==3.8.1
boto3==1.34.144
botocore==1.34.144
certifi==2024.6.2
charset-normalizer==3.3.2
decorator==4.4.2
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.1.0
gunicorn==22.0.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
inflection==0.5.1
jmespath==1.0.1
mccabe==0.7.0
moviepy==1.0.3
numpy==2.0.0
packaging==24.1
pillow==10.3.0
proglog==0.1.10
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2
setuptools==70.1.1
six==1.16.0
sqlparse==0.5.0
tqdm==4.66.4
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
